move netfx failing tests to String.netcoreapp.cs
