+++ src/Common/tests/Configurations.props
@@ -3,8 +3,7 @@
   <PropertyGroup>
     <BuildConfigurations>
       netcoreapp-Unix;
-      netcoreapp-Windows_NT;
-      netfx;
+      netcoreapp-Windows_NT;      
     </BuildConfigurations>
   </PropertyGroup>
 </Project>
