+++ src/System.Memory/tests/System.Memory.Tests.csproj
@@ -160,7 +160,12 @@
     </Compile>
     <Compile Include="$(CommonTestPath)\Tests\System\StringTests.cs">
       <Link>ReadOnlySpan\StringTests.cs</Link>
-    </Compile>    
+    </Compile>
+  </ItemGroup>
+  <ItemGroup Condition="'$(TargetGroup)' != 'netfx'">
+    <Compile Include="$(CommonTestPath)\Tests\System\StringTests.netcoreapp.cs">
+      <Link>ReadOnlySpan\StringTests.netcoreapp.cs</Link>
+    </Compile>
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Memory\AsMemory.cs" />
